requirements.txt
aiohttp==3.9.1
aiosqlite>=0.21.0
discord-py>=2.5.2
email-validator>=2.2.0
flask>=3.1.0
flask-sqlalchemy>=3.1.1
google-generativeai==0.3.2
gunicorn>=23.0.0
newsapi>=0.1.1
newsapi-python>=0.2.7
psycopg2-binary>=2.9.10
requests>=2.32.3
trafilatura>=2.0.0